+++ packages/reactabular-tree/tests/unpack_test.js
@@ -2,167 +2,96 @@ import { expect } from 'chai';
 import { unpack } from '../src';
 
 describe('tree.unpack', function () {
-  it('returns an empty array if empty rows are passed', function () {
-    expect(unpack([])).to.deep.equal([]);
+  it('returns an empty tree as an empty array', function () {
+    expect(unpack()([])).to.deep.equal([]);
   });
 
-  it('returns the array itself if there is nothing to unpack', function () {
-    const given = [
+  it('returns an unpacked array', function () {
+    const rows = [
       {
-        foo: 'bar'
-      }
-    ];
-
-    expect(unpack(given)).to.deep.equal(given);
-  });
-
-  it('unpacks a single child', function () {
-    const given = [
-      {
-        id: 0,
-        foo: 'bar',
-        _pack: [
+        id: 'foobar',
+        name: 'foo',
+        children: [
           {
-            parent: 0,
-            foo: 'foo'
+            id: 'barfoo',
+            name: 'bar'
           }
         ]
       }
     ];
     const expected = [
       {
-        id: 0,
-        foo: 'bar'
+        id: 'foobar',
+        name: 'foo'
       },
       {
-        parent: 0,
-        foo: 'foo'
+        parent: 'foobar',
+        id: 'barfoo',
+        name: 'bar'
       }
     ];
 
-    expect(unpack(given)).to.deep.equal(expected);
+    expect(unpack()(rows)).to.deep.equal(expected);
   });
 
-  it('works with sibling children', function () {
-    const given = [
-      {
-        id: 0,
-        foo: 'bar',
-        _pack: [
-          {
-            id: 1,
-            parent: 0,
-            foo: 'foo'
-          },
-          {
-            id: 2,
-            parent: 0,
-            foo: 'barbar'
-          }
-        ]
-      }
-    ];
-    const expected = [
-      {
-        id: 0,
-        foo: 'bar'
-      },
-      {
-        id: 1,
-        parent: 0,
-        foo: 'foo'
-      },
-      {
-        id: 2,
-        parent: 0,
-        foo: 'barbar'
-      }
-    ];
-
-    expect(unpack(given)).to.deep.equal(expected);
-  });
-
-  it('works with nested children', function () {
-    // TODO: this could be made recursive
-    const given = [
-      {
-        id: 0,
-        foo: 'bar',
-        _pack: [
-          {
-            id: 1,
-            parent: 0,
-            foo: 'foo'
-          },
-          {
-            id: 2,
-            parent: 1,
-            foo: 'barbar'
-          }
-        ]
-      }
-    ];
-    const expected = [
-      {
-        id: 0,
-        foo: 'bar'
-      },
-      {
-        id: 1,
-        parent: 0,
-        foo: 'foo'
-      },
-      {
-        id: 2,
-        parent: 1,
-        foo: 'barbar'
-      }
-    ];
-
-    expect(unpack(given)).to.deep.equal(expected);
-  });
-
-  // TODO
-  /*
   it('allows parent field to be customized', function () {
-    const property = 'demo';
-    // TODO: this could be made recursive
-    const given = [
+    const rows = [
       {
-        id: 0,
-        foo: 'bar',
-        _pack: [
+        id: 'foobar',
+        name: 'foo',
+        children: [
           {
-            id: 1,
-            [property]: 0,
-            foo: 'foo'
-          },
-          {
-            id: 2,
-            [property]: 1,
-            foo: 'barbar'
+            id: 'barfoo',
+            name: 'bar'
           }
         ]
       }
     ];
     const expected = [
       {
-        id: 0,
-        foo: 'bar'
+        id: 'foobar',
+        name: 'foo'
       },
       {
-        id: 1,
-        [property]: 0,
-        foo: 'foo'
-      },
-      {
-        id: 2,
-        [property]: 1,
-        foo: 'barbar'
+        p: 'foobar',
+        id: 'barfoo',
+        name: 'bar'
       }
     ];
 
-    expect(unpack({ parent: property })(given)).to.deep.equal(expected);
+    expect(unpack({ parentField: 'p' })(rows)).to.deep.equal(expected);
   });
-  */
+
+
+  it('allows id field to be customized', function () {
+    const rows = [
+      {
+        _id: 'foobar',
+        name: 'foo',
+        children: [
+          {
+            _id: 'barfoo',
+            name: 'bar'
+          }
+        ]
+      }
+    ];
+    const expected = [
+      {
+        _id: 'foobar',
+        name: 'foo'
+      },
+      {
+        parent: 'foobar',
+        _id: 'barfoo',
+        name: 'bar'
+      }
+    ];
+
+    expect(unpack({ idField: '_id' })(rows)).to.deep.equal(expected);
+  });
+
+  it('returns an empty array with invalid input', function () {
+    expect(unpack()('foobar')).to.deep.equal([]);
+  });
 });